docs: add note about virtcontainers if testing runtime
If following the contributing doc and wanting to run make check and
make test in kata-runtime, the virtcontainers/bin/hook executable
is expected to exist. But it only exists if it is compiled and
installed. So a note has been added that if make test|check for
runtime, to ensure that virtcontainers has been compiled and
installed.

Signed-off-by: ccollicutt <curtis@serverascode.com>