Specifying timeline for defining the electorate
This patch specifies the first day of the nomination period as the last day
for anyone to qualify as eligible electorate.

This will help with building the electorate list earlier in the process,
so election officials can make sure that everyone is included before
creating the poll for the voting period.

Signed-off-by: Ildiko Vancsa <ildiko.vancsa@gmail.com>