readme/elections: Fabiano Fidêncio is stepping down
After 3.5 years serving the community as part of the Architecture
Committee, this is the time for me to step down, and save some time to
focus on specific areas of the project that I currently am lacking the
time to do so.

It's been an amazing ride, I'm thankful for all we could achieve
together, and I won't be a strange in the community.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>