ci: Rework static checks
The static checks were using the version from kata-containers,
an old version of go and had a bunch of travis references that
don't seem to be required any longer, so try updating this
and removing the redundant code.

Whether all these checks are useful is debatable, but at least
now they should be up-to-date.

Signed-off-by: stevenhorsman <steven@uk.ibm.com>